CogniPath/CogniPathLMS/server/requirements.txt
flask
flask-cors
google-genai
python-dotenv
pypdf
python-docx
gunicorn
firebase-admin